Migrated CountdownTimer.
